Basic Terrain Generation
Unity 2018.2 Required

These assets are used for generating a small landscape. Terrain generation is multithreaded. I attempted to add ambient occlusion and light propagation in this revision, but I ran into an issue with light propagating into chunks that did not exist or already had a mesh. This will be resolved in a later build.